refactor: readability uses snapshot instead of link
